Refactor of flexible_type <--> python code.
This pull request reworks the flexible_type translation code.  New features:

1. Speedup: The new code is between 10% and 4x faster in constructing
an SArray/SFrame, depending on type, and 2-4x faster in accessing
SArray/SFrame elements.  A future work is to apply these changes
to the pylambda workers.

2. Consistent type inference rules.  Type inference is now handled
consistently when multiple types are present, with rules that
cover all the corner cases. Previously, many of these corner cases
were not handled correcly.

- Integers, floats, strings, datetime, dict, arrays, and image all
have one-to-one correspondence with a flexible_type.

- Lists are translated to a vector if all elements are numeric, otherwise
they are translated to a list.

- An empty list is treated as a list.

- None is translated to flex_undefined.

Now, when type inference is done, and multiple types appear in an input
(e.g. empty lists and numeric lists in the input to an SArray), then the
following rules goven what type changes can be made in inferring a common
type:

- Rule 1: Integer can be upgraded to a float.
- Rule 2: A numeric list and a vector together make a vector
- Rule 3: None does not affect the type.  I.e. [t1,t2,None, t3] has the same
infered type as [t1, t2, t3]
- Rule 4: If a list or vector is present, an empty list does not
change the infered type.
- Rule 5: A vector can be upgraded to a list.

If a type cannot be chosen that is consistent with these rules, then a TypeError
exception is raised.

3. Type casting -- now the behavior when calling sf.SArray(data, dtype = type)
follows the same rules as SArray.astype.  Previously, this was part way
between type inference (very strict) and astype(...) loose, allowing more type
casting but basically following its own rules.  The behavior is now consistent
with astype.