Bugfix: ODBC maximum string length calculation
Before writing a table to a SQL database, we must find the max length
for each column to know how much memory to allocate for a row. Our code
to do this for string and image type columns was slightly wrong, as it
stopped calculating the maximum if it ever found a NULL in the column.
So whatever was the maximum so far in that segment was taken to be the
maximum. This caused an infrequent error in certain datasets. Fixed
with this.