Bugfix: Mishandled NULL in ODBC connector
Since we store "time intervals" as dictionaries because we don't have a
time interval type in SFrame, but ODBC does not support list or
dictionary-like types, we need to check a dictionary column to see if it
is intervals (legal) or other random dicts (illegal). This check would
break if there was a NULL in the column.