Improve logging when launching processes
When launching a process (unity_server or pylambda_worker in this
commit) make sure that if the process errors out, the error message is
delivered to the user in some form (either to terminal or log). Also
some commands in the top namespace to print the location of the log to
look for errors.

Testing was in the form of manual adversarial tests, as it involved
breaking a current installation by moving files around, or changing
other system stuff.