Fix failure to find supporting DLLs on Win32
Since we are launching the pylambda_worker as a DLL from python.exe, we can no
longer expect the supporting DLLs we package within the egg to be found by it.
This augments the path of unity_server on Windows only to add the graphlab path
within site-packages. This environment is inherited by children.