Fix set_windows_dll_path
The current use and implementation of set_windows_dll_path
is problematic in several ways:
1. The intended behavior of set_windows_dll_path
is to add "libpylambda" to the dll path. The current implementation
replaces the entire search path. This causes future loadlibrary
to fail.
2. the function is called inside make_unity_server_env() which
should not have side effects.

This fix makes the following changes:
1. Rename set_windows_dll_path() to add_pylambda_windows_dll_path()
2. Use AddDllDirectory instead of SetDllDirectory.
3. Move the function call out of make_unity_server_evn() and call
it explicitly on server start