pyspark_unity now only needs 2 .so files.
It appears that (a) the python include pxi file does not do what it
says it does, and (b) cimported cython files still get issued a python
include statement, even if they are not actually pulling in any python
files. so it works fine to include multiple .pyx source files, but
then the executable still issues a python-level import command
However, what we can do is just copy the 2 libs to the executable
directory, and then it should work fine.