Add "use_exact_column_names" option to to_sql.
This changes the default behavior of to_sql. to_sql will now submit data
expecting the column names in the SFrame to match up exactly to the
column names of the table it is inserting into. If this option is set to
False, then the old behavior will take over, where the data values in
the SFrame are simply submitted in order to the table without care of
what columns are in the table.