Critical Bug fix to logical filter
The following code fails due to the inability to prove that a and a2 have the
same length when performing the logical filter. The code is supposed to
selectively materialize the filter so that length can be proven. However,
due to a bug in the prove_equal_length method, that was not done.

```
g=SArray(range(10000))
g2=SArray(range(10000))
a=g[g>10]
a2=g2[g>10]
a[a2>20] # Fail here
```