Update type determination in from_sql
Formalizes and documents the steps taken to determine the column types
of an SFrame given a result set from a DBAPI2 cursor. Prior to this,
column_type_hints didn't work as documented, and now results in a
Python-side cast to the hinted types. This also tries to make a best
effort not to fail on type conversion, and adds a check to see if a
returned type is actually supported by SFrame, falling back to string if
it's not.