Fixes long file name issue with ipc:// addresses
ipc:// addresses can fail if the filename exceeds 104 characters
on Mac, and 108 characters on Linux. This is a limitation of Unix Domain
Sockets.  See http://man7.org/linux/man-pages/man7/unix.7.html for details.

Here we add a preprocess that if the filename exceeds this length, we
switch it to "/tmp/" + md5hash(address)

Also fixed some warnings on clang. (Warning move prevents copy elision).