Fix memory leak in png_encode
The image_type uses boost::shared_ptr<char[]> to manage the image bytearray
in which the pointer is required to be allocated using C++ new char[].

In the png encoder, we use malloc/realloc. This fixes the leak by
makeing a copy of the png encoded memory buffer into a "new" allocated
array.